# utils: array-search: optimize array_search_gt