clang's autovectorizer does a better job than our handwritten SIMD.

`perf_collection_g --test find --col bptree`, with node size 4

Before: fill: 2092.890381 ms, find: 1687.587402 ms
After: fill: 2022.945679 ms, find: 1520.288330 ms

`perf_collection_g --test find --col bptree`, with node size 16

Before: fill: 2163.744629 ms, find: 1789.123901 ms
After: fill: 2065.735352 ms find: 1523.582153 ms

Those are results from my laptop, but on AWS (m5d) the autovectorized
version is faster too.