@avikivity 2021-08-10
/cc @xemul . Can you show generated code?