@avikivity 2021-08-10
Couldn't resist and godbolted it:

https://godbolt.org/z/4WE7W6c5z