@michoecho 2021-08-10
Comparison of both versions: https://godbolt.org/z/Pfa9PadYr