@avikivity 2021-08-10
The generated code doesn't look very impressive, and it also can't assume that capacity is a multiple of 4, yet it's faster. I guess movemask (which moved between the vector and int domains) and popcnt are extra slow. Maybe summing in the vector domain, then doing a horizontal add, then 2x shuffle + add would be faster.