@avikivity 2021-08-10
vpmovmskb is actually just 3-4 cycles. I don't get it.