@michoecho 2021-08-10
> Maybe it's because the hand-written version is unrolled sequentially, and the compiler version runs two iterations in parallel.

The handwritten version is 2-interleaved, but the autovectorized version is 4-interleaved.