@avikivity 2021-08-10
Clearly the compiler is giving its own code an unfair advantage.

4-way unroll means just one iteration of the loop will run.