@michoecho 2021-08-10
> just 3-4 cycles

What do you mean by "just" 4 cycles? Current intel CPUs have 3 Vect ALUs, so you can do 36 arithmetic ops (64-bit) in those 4 cycles.