@avikivity 2021-08-10
Just loading a vector is 5 cycles, best case (hit in L1): https://uops.info/html-instr/VLDDQU_YMM_M256.html

Most simple ops have 1 cycle latency.

Those 3 cycles shouldn't affect anything, since the CPU can start the next iteration in parallel with this computation. I guess we're seeing the difference between 4-way and 2-way unroll, with the better interleaving of 4-way unroll. I also think that you'll see very little difference in perf_simple_query which will dilute the improvement greatly. Of course, the improvement is still worthwhile, especially when it is achieved by deleting code.