@avikivity 2021-08-10
vextracti128, which the compiler uses so liberally, is also 3 cycles: https://uops.info/html-instr/VEXTRACTI128_XMM_YMM_I8.html